conveyorguard-api/requirements.txt
# ConveyorGuard ML API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
torch>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Supabase + Auth
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.24.0
python-dotenv>=1.0.0
